Add AAB icon parsing and share impl with APK icon parsing
